# Optimize volume integral kernels for flux differencing