--- a/benchmark/advection_basic_2d.jl
+++ b/benchmark/advection_basic_2d.jl
@@ -37,8 +37,10 @@ source_terms_gpu = semi_gpu.source_terms
 
 # ODE on GPU
 ode_gpu = semidiscretizeGPU(semi_gpu, tspan_gpu)
-u_gpu = copy(ode_gpu.u0)
-du_gpu = similar(u_gpu)
+u_gpu_ = copy(ode_gpu.u0)
+du_gpu_ = similar(u_gpu_)
+u_gpu = TrixiCUDA.wrap_array(u_gpu_, mesh_gpu, equations_gpu, solver_gpu, cache_gpu)
+du_gpu = TrixiCUDA.wrap_array(du_gpu_, mesh_gpu, equations_gpu, solver_gpu, cache_gpu)
 
 # Reset du and volume integral
 @info "Time for reset_du! and volume_integral! on GPU"