--- a/src/solvers/dg_1d.jl
+++ b/src/solvers/dg_1d.jl
@@ -236,7 +236,71 @@ end
 ############################################################################## New optimization
 # Kernel for calculating symmetric and nonconservative volume fluxes and 
 # corresponding volume integrals
-function noncons_volume_flux_integral_kernel!()
+function noncons_volume_flux_integral_kernel!(du, u, derivative_split, derivative_split_zero,
+                                              equations::AbstractEquations{1},
+                                              symmetric_flux::Any, nonconservative_flux::Any)
+    # Set tile width
+    tile_width = size(du, 2)
+    offset = 0 # offset bytes for shared memory
+
+    # Allocate dynamic shared memory
+    shmem_split = @cuDynamicSharedMem(eltype(du), (tile_width, tile_width))
+    offset += sizeof(eltype(du)) * tile_width^2
+    shmem_szero = @cuDynamicSharedMem(eltype(du), (tile_width, tile_width),
+                                      offset)
+    offset += sizeof(eltype(du)) * tile_width^2
+    shmem_value = @cuDynamicSharedMem(eltype(du), (size(du, 1), tile_width),
+                                      offset)
+
+    # Get thread and block indices only we need to save registers
+    ty = threadIdx().y
+    k = (blockIdx().z - 1) * blockDim().z + threadIdx().z
+
+    # Tile the computation (set to one tile here)
+    # Initialize the values
+    for tx in axes(du, 1)
+        @inbounds shmem_value[tx, ty] = zero(eltype(du))
+    end
+
+    # Load data from global memory into shared memory
+    for ty2 in axes(du, 2)
+        # Transposed access
+        @inbounds begin
+            shmem_split[ty2, ty] = derivative_split[ty, ty2]
+            shmem_szero[ty2, ty] = derivative_split_zero[ty, ty2]
+        end
+    end
+
+    # Compute volume fluxes
+    # How to store nodes in shared memory?
+    for thread in 1:tile_width
+        # Volume flux is heavy in computation so we should try best to avoid redundant 
+        # computation, i.e., use for loop along x direction here
+        u_node = get_node_vars(u, equations, ty, k)
+        symmetric_flux_node = symmetric_flux(u_node,
+                                             get_node_vars(u, equations, thread, k),
+                                             1, equations)
+        noncons_flux_node = nonconservative_flux(u_node,
+                                                 get_node_vars(u, equations, thread, k),
+                                                 1, equations)
+
+        # TODO: Avoid potential bank conflicts
+        for tx in axes(du, 1)
+            @inbounds begin
+                shmem_value[tx, ty] += symmetric_flux_node[tx] * shmem_szero[ty, thread]
+                shmem_value[tx, ty] += 0.5f0 * noncons_flux_node[tx] * shmem_split[ty, thread]
+            end
+        end
+    end
+
+    # Synchronization is not needed here if we use only one tile
+    # sync_threads()
+
+    # Finalize the values
+    for tx in axes(du, 1)
+        @inbounds du[tx, ty, k] = shmem_value[tx, ty]
+    end
+
     return nothing
 end
 
@@ -721,8 +785,9 @@ function cuda_volume_integral!(du, u, mesh::TreeMesh{1}, nonconservative_terms,
     # method. However, there are other factors that may cause a launch failure, such as the maximum 
     # shared memory per block. Here, we have omitted all other factors, but this should be enhanced 
     # later for a safer kernel launch.
+
     # TODO: More checks before the kernel launch
-    thread_per_block = size(du, 1) * size(du, 2)^2
+    thread_per_block = size(du, 1) * size(du, 2)
     if thread_per_block > MAX_THREADS_PER_BLOCK
         # TODO: How to optimize when size is large
         flux_arr = similar(u)
@@ -755,13 +820,12 @@ function cuda_volume_integral!(du, u, mesh::TreeMesh{1}, nonconservative_terms::
 
     volume_flux = volume_integral.volume_flux
 
+    # TODO: Move `set_diagonal_to_zero!` outside of loop and cache the result in DG on GPU
     derivative_split = dg.basis.derivative_split
-    # TODO: Move `set_diagonal_to_zero!` outside of `rhs!` loop and cache the result in 
-    # DG struct on GPU
     set_diagonal_to_zero!(derivative_split)
     derivative_split = CuArray(derivative_split)
 
-    thread_per_block = size(du, 1) * size(du, 2)^2
+    thread_per_block = size(du, 2)
     if thread_per_block > MAX_THREADS_PER_BLOCK
         # TODO: How to optimize when size is large
         volume_flux_arr = CuArray{RealT}(undef, size(u, 1), size(u, 2), size(u, 2), size(u, 3))
@@ -795,32 +859,48 @@ function cuda_volume_integral!(du, u, mesh::TreeMesh{1}, nonconservative_terms::
 
     symmetric_flux, nonconservative_flux = dg.volume_integral.volume_flux
 
-    derivative_split = dg.basis.derivative_split
-    set_diagonal_to_zero!(derivative_split) # temporarily set here, maybe move outside `rhs!`
+    # TODO: Move `set_diagonal_to_zero!` outside of loop and cache the result in DG on GPU
+    derivative_split = dg.basis.derivative_split # create copy
+    set_diagonal_to_zero!(derivative_split)
+    derivative_split_zero = CuArray(derivative_split)
+    derivative_split = CuArray(dg.basis.derivative_split)
 
-    derivative_split = CuArray(derivative_split)
-    symmetric_flux_arr = CuArray{RealT}(undef, size(u, 1), size(u, 2), size(u, 2), size(u, 3))
-    noncons_flux_arr = CuArray{RealT}(undef, size(u, 1), size(u, 2), size(u, 2), size(u, 3))
+    thread_per_block = size(du, 2)
+    if thread_per_block > MAX_THREADS_PER_BLOCK
+        # TODO: How to optimize when size is large
+        symmetric_flux_arr = CuArray{RealT}(undef, size(u, 1), size(u, 2), size(u, 2), size(u, 3))
+        noncons_flux_arr = CuArray{RealT}(undef, size(u, 1), size(u, 2), size(u, 2), size(u, 3))
 
-    noncons_volume_flux_kernel = @cuda launch=false noncons_volume_flux_kernel!(symmetric_flux_arr,
-                                                                                noncons_flux_arr,
-                                                                                u,
-                                                                                derivative_split,
-                                                                                equations,
-                                                                                symmetric_flux,
-                                                                                nonconservative_flux)
-    noncons_volume_flux_kernel(symmetric_flux_arr, noncons_flux_arr, u, derivative_split,
-                               equations, symmetric_flux, nonconservative_flux;
-                               kernel_configurator_2d(noncons_volume_flux_kernel,
-                                                      size(u, 2)^2, size(u, 3))...)
+        noncons_volume_flux_kernel = @cuda launch=false noncons_volume_flux_kernel!(symmetric_flux_arr,
+                                                                                    noncons_flux_arr, u,
+                                                                                    derivative_split_zero,
+                                                                                    equations,
+                                                                                    symmetric_flux,
+                                                                                    nonconservative_flux)
+        noncons_volume_flux_kernel(symmetric_flux_arr, noncons_flux_arr, u, derivative_split_zero,
+                                   equations, symmetric_flux, nonconservative_flux;
+                                   kernel_configurator_2d(noncons_volume_flux_kernel,
+                                                          size(u, 2)^2, size(u, 3))...)
 
-    derivative_split = CuArray(dg.basis.derivative_split) # use original `derivative_split`
-
-    volume_integral_kernel = @cuda launch=false volume_integral_kernel!(du, derivative_split,
-                                                                        symmetric_flux_arr,
-                                                                        noncons_flux_arr)
-    volume_integral_kernel(du, derivative_split, symmetric_flux_arr, noncons_flux_arr;
-                           kernel_configurator_3d(volume_integral_kernel, size(du)...)...)
+        volume_integral_kernel = @cuda launch=false volume_integral_kernel!(du, derivative_split,
+                                                                            symmetric_flux_arr,
+                                                                            noncons_flux_arr)
+        volume_integral_kernel(du, derivative_split, symmetric_flux_arr, noncons_flux_arr;
+                               kernel_configurator_3d(volume_integral_kernel, size(du)...)...)
+    else
+        shmem_size = (size(du, 2)^2 * 2 + size(du, 1) * size(du, 2)) * sizeof(eltype(du))
+        noncons_volume_flux_integral_kernel = @cuda launch=false noncons_volume_flux_integral_kernel!(du, u,
+                                                                                                      derivative_split,
+                                                                                                      derivative_split_zero,
+                                                                                                      equations,
+                                                                                                      symmetric_flux,
+                                                                                                      nonconservative_flux)
+        noncons_volume_flux_integral_kernel(du, u, derivative_split, derivative_split_zero, equations,
+                                            symmetric_flux, nonconservative_flux;
+                                            shemem = shmem_size,
+                                            threads = (1, size(du, 2), 1),
+                                            blocks = (1, 1, size(du, 3)))
+    end
 
     return nothing
 end