Continue #105 with a focus on `nonconservative_terms::True`.