@huiyuxie 2025-01-14
We choose mhd with volume flux differencing for benchmarks.