@huiyuxie 2025-01-14
The optimization is not really good for case 1D.

Before the optimization
```
[ Info: Time for reset_du! and volume_integral! on GPU
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  18.900 μs … 871.100 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     20.300 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   25.552 μs ±  32.783 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▇█▆▃▄▂▁                                                      ▁
  █████████▇██▇▇██▇▇▆▆▆█▇█▇▇▅▆▅▇█▆▆▆▅▆▇▇▆▆▇▇▇▇▅▅▅▄▄▄▃▄▆▂▇▇▆▆▆▆ █
  18.9 μs       Histogram: log(frequency) by time        77 μs <

 Memory estimate: 1.88 KiB, allocs estimate: 32.
```
After the optimization
```
[ Info: Time for reset_du! and volume_integral! on GPU
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  19.700 μs … 202.600 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     22.000 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   26.193 μs ±  12.016 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▃▇█▅▂▂▄▅▄▂▁                                                  ▁
  ███████████▇▇▆█▇▇▇▆█▇▆█▇█▇▇▅▆▇▇▇▆▅▆▇▆▇▅▆▆▆▆▆▆▆▆▅▅▆▆▆▆▅▆▆▇▆▇▇ █
  19.7 μs       Histogram: log(frequency) by time      77.1 μs <

 Memory estimate: 1.88 KiB, allocs estimate: 32.
```