@huiyuxie 2025-01-15
The optimization works well for case 2D.

Before the optimization
```
[ Info: Time for reset_du! and volume_integral! on GPU
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  106.300 μs …  28.088 ms  ┊ GC (min … max): 0.00% … 19.55%
 Time  (median):     112.200 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   126.037 μs ± 430.446 μs  ┊ GC (mean ± σ):  1.16% ±  0.34%

    ▆█
  ▁▅███▃▂▂▂▂▂▃▄▅▅▅▃▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  106 μs           Histogram: frequency by time          192 μs <

 Memory estimate: 6.58 KiB, allocs estimate: 102.
```
After the optimization
```
[ Info: Time for reset_du! and volume_integral! on GPU
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  35.700 μs … 278.900 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     39.300 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   49.352 μs ±  19.341 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▂▅█▅▃▁▁        ▁▂▂▂▃▁   ▂▄▃▂ ▁▁                       ▂▁     ▁
  ████████▇▇▆▆▇████████▇▇██████████▆▆▅▅▅▂▄▄▃▅▅▂▂▃▃▃▄▃▄▆█████▆▆ █
  35.7 μs       Histogram: log(frequency) by time       110 μs <

 Memory estimate: 2.45 KiB, allocs estimate: 42.
```