@huiyuxie 2025-01-15
The optimization also works well for case 3D.

Before the optimization
```
[ Info: Time for reset_du! and volume_integral! on GPU
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  166.000 μs …  21.309 ms  ┊ GC (min … max): 0.00% … 18.31%
 Time  (median):     207.700 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   223.133 μs ± 412.949 μs  ┊ GC (mean ± σ):  0.68% ±  0.36%

             ▃█▆▁▅▅▃▁
  ▂▂▂▂▁▁▁▂▁▃▇████████▇▇▇▇▅▄▄▃▃▄▃▃▃▃▃▃▃▃▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ ▃
  166 μs           Histogram: frequency by time          320 μs <

 Memory estimate: 8.58 KiB, allocs estimate: 124.
```
After the optimization
```
[ Info: Time for reset_du! and volume_integral! on GPU
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  41.700 μs … 260.800 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     44.900 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   47.063 μs ±   9.970 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▄▅▇█▇▄▃▂▁                                                   ▂
  ███████████▇▇▅▆▅▅▄▄▄▅▅▆███▇▇▇▇██▇█████▇▆▅▆▅▄▅▅▄▅▅▄▄▅▄▄▂▄▅▃▄▄ █
  41.7 μs       Histogram: log(frequency) by time      82.8 μs <

 Memory estimate: 2.52 KiB, allocs estimate: 41.
```